refactor: integrate ValidateCodeCountdown for improved countdown management and streamline resend logic
