+++ src/pages/signin/ValidateCodeCountdown/index.tsx
@@ -1,18 +1,10 @@
 import React, {forwardRef, useEffect, useImperativeHandle, useRef, useState} from 'react';
-import {View} from 'react-native';
-import PressableWithFeedback from '@components/Pressable/PressableWithFeedback';
 import RenderHTML from '@components/RenderHTML';
-import Text from '@components/Text';
 import useLocalize from '@hooks/useLocalize';
-import useStyleUtils from '@hooks/useStyleUtils';
-import useThemeStyles from '@hooks/useThemeStyles';
 import CONST from '@src/CONST';
 import type {ValidateCodeCountdownHandle, ValidateCodeCountdownProps} from './types';
 
-const ValidateCodeCountdown = forwardRef<ValidateCodeCountdownHandle, ValidateCodeCountdownProps>(({shouldDisableResendValidateCode, hasError, isOffline, onResendValidateCode}, ref) => {
-    const StyleUtils = useStyleUtils();
-    const styles = useThemeStyles();
-
+const ValidateCodeCountdown = forwardRef<ValidateCodeCountdownHandle, ValidateCodeCountdownProps>(({onCountdownFinish}, ref) => {
     const {translate} = useLocalize();
 
     const [timeRemaining, setTimeRemaining] = useState<number>(CONST.REQUEST_CODE_DELAY);
@@ -27,38 +19,20 @@ const ValidateCodeCountdown = forwardRef<ValidateCodeCountdownHandle, ValidateCo
             timerRef.current = setTimeout(() => {
                 setTimeRemaining((prev) => prev - 1);
             }, 1000);
+        } else {
+            onCountdownFinish();
         }
         return () => {
             clearTimeout(timerRef.current);
         };
-    }, [timeRemaining]);
+    }, [onCountdownFinish, timeRemaining]);
 
     return (
-        <View style={[styles.alignItemsStart]}>
-            {timeRemaining > 0 && !isOffline ? (
-                <View style={[styles.mt2, styles.flexRow, styles.renderHTML]}>
-                    <RenderHTML
-                        html={translate('validateCodeForm.requestNewCode', {
-                            timeRemaining: `00:${String(timeRemaining).padStart(2, '0')}`,
-                        })}
-                    />
-                </View>
-            ) : (
-                <PressableWithFeedback
-                    style={[styles.mt2]}
-                    onPress={onResendValidateCode}
-                    disabled={shouldDisableResendValidateCode}
-                    hoverDimmingValue={1}
-                    pressDimmingValue={0.2}
-                    role={CONST.ROLE.BUTTON}
-                    accessibilityLabel={translate('validateCodeForm.magicCodeNotReceived')}
-                >
-                    <Text style={[StyleUtils.getDisabledLinkStyles(shouldDisableResendValidateCode)]}>
-                        {hasError ? translate('validateCodeForm.requestNewCodeAfterErrorOccurred') : translate('validateCodeForm.magicCodeNotReceived')}
-                    </Text>
-                </PressableWithFeedback>
-            )}
-        </View>
+        <RenderHTML
+            html={translate('validateCodeForm.requestNewCode', {
+                timeRemaining: `00:${String(timeRemaining).padStart(2, '0')}`,
+            })}
+        />
     );
 });
